refactor: 도메인내 클래스 패키지 분리
