+++ src/main/java/blackjack/controller/GameController.java
@@ -1,10 +1,10 @@
 package blackjack.controller;
 
-import blackjack.domain.Dealer;
-import blackjack.domain.Deck;
+import blackjack.domain.participant.Dealer;
+import blackjack.domain.deck.Deck;
 import blackjack.domain.Game;
-import blackjack.domain.Player;
-import blackjack.domain.Players;
+import blackjack.domain.participant.Player;
+import blackjack.domain.participant.Players;
 import blackjack.view.InputView;
 import blackjack.view.OutputView;
