+++ src/main/java/blackjack/domain/deck/Deck.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.deck;
 
 import java.util.ArrayDeque;
 import java.util.Collections;
