+++ src/main/java/blackjack/domain/participant/Dealer.java
@@ -1,4 +1,6 @@
-package blackjack.domain;
+package blackjack.domain.participant;
+
+import blackjack.domain.deck.Deck;
 
 public class Dealer extends Participant {
