+++ src/test/java/blackjack/domain/DealerTest.java
@@ -1,5 +1,8 @@
 package blackjack.domain;
 
+import blackjack.domain.deck.Card;
+import blackjack.domain.deck.Deck;
+import blackjack.domain.participant.Dealer;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
