+++ src/test/java/blackjack/domain/DeckTest.java
@@ -1,5 +1,7 @@
 package blackjack.domain;
 
+import blackjack.domain.deck.Card;
+import blackjack.domain.deck.Deck;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
