+++ src/test/java/blackjack/domain/HandsScoreTest.java
@@ -1,5 +1,6 @@
 package blackjack.domain;
 
+import blackjack.domain.participant.HandsScore;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
