+++ src/test/java/blackjack/domain/ValueTest.java
@@ -1,11 +1,12 @@
 package blackjack.domain;
 
+import blackjack.domain.deck.Value;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import static blackjack.domain.Value.findValue;
+import static blackjack.domain.deck.Value.findValue;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
